pin-1385: Move all projections under the same package
